Enable erasing precision directives from variable declarations based on earlier precision declarations or default precision based on source type. Export attribute and uniform name information for GLSL source when those attributes or uniforms have no layout directives.

+++ src/quad_glesv2.vert.glsl
@@ -1,6 +1,6 @@
 attribute vec2 vertex;
 
-varying mediump vec2 position;
+varying highp vec2 position;
 
 void main()
 {
